Create practice2.py
This file a contains a simple password generator, which generates a random password from 8 characters to 32 characters in length. This code sample uses the string library and utilises some of its features like, .join, choice, string.ascii_letters, string.punctuaion, string.digits. It also uses the random library to decide the length of the password